requirements.txt
numpy>=1.20.3
six>=1.15.0
torch>=1.10.0
torchvision>=0.11.1
pillow>=8.4.0,<=9.0.0
